Moved synchronization to lib
