+++ src/lib/synchronization/CommitFinder.js
